requirements.txt
pycryptodome
pkcs7
jinja2
termcolor
Flask
